+++ bin/blocke-api.js
@@ -6,9 +6,19 @@ const OptionRequestHandler = require('../lib/option-request-handler');
 const version = require('../package.json').version;
 const _ = require('underscore');
 
+module.exports = {
+    executeHandlerApi: function (coin, options, callback) {
+        executeHandler(coin, options, callback);
+    }
+}
 
+function executeHandler(coin, options, callback) {
 
-function executeHandler(handler, usage, callback) {
+    var command = coin;
+    shortHandCommand = command;
+
+    const handler = new OptionRequestHandler(shortHandCommand, options);
+    const usage = commandLineUsage(commandLineArgData[command].usage);
 
     handler.handleRequest().then((res) => {
